requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-debug-toolbar==4.2.0
drf-spectacular==0.27.0
psycopg2-binary>=2.9.9
redis==5.0.1
PyJWT==2.8.0
celery==5.3.4
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.1.0
django-storages==1.14.2
boto3==1.34.0
psutil==5.9.6

# Мониторинг и метрики
django-prometheus==2.3.1
prometheus-client==0.19.0

# SSL и безопасность
django-extensions==3.2.3
cryptography==41.0.7 